Cast away warning
Not yet fixed upstream.

Review URL: https://codereview.chromium.org/167973004

git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/libvpx@251547 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

+++ source/libvpx/vp9/vp9_dx_iface.c
@@ -214,7 +214,7 @@ static vpx_codec_err_t vp9_get_si(vpx_codec_alg_priv_t    *ctx,
                        ? sizeof(vp9_stream_info_t)
                        : sizeof(vpx_codec_stream_info_t);
   memcpy(si, &ctx->si, sz);
-  si->sz = sz;
+  si->sz = (unsigned int)sz;
 
   return VPX_CODEC_OK;
 }
